#65 | Combined clock again
